connect all styles, distribute it to different files
