+++ package.json
@@ -2,6 +2,7 @@
   "name": "test-task",
   "version": "1.0.0",
   "dependencies": {
+    "color": "^3.1.3",
     "react": "^17.0.1",
     "react-dom": "^17.0.1"
   },
