+++ src/assets/styles/container-style.js
@@ -1,0 +1,43 @@
+import { createStyles, makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => createStyles({
+  '@font-face': {
+    fontFamily: 'Montserrat',
+    src: 'url(../fonts/Montserrat-Regular.ttf)',
+  },
+  h1class: {
+    width: '50%',
+    textAlign: 'center',
+    fontSize: '2em',
+    fontWeight: 900,
+    color: 'white',
+    border: '1px solid white',
+    borderRadius: '5px',
+  },
+  todoContainer: {
+    maxWidth: '80%',
+    width: '50%',
+    minWidth: '50%',
+    margin: '1em auto 0',
+    padding: '10px 50px',
+    background: 'darkcyan',
+    boxShadow: '0 0 25px deepskyblue',
+    border: 'none',
+    boxSizing: 'border-box',
+    fontFamily: 'Montserrat, sans-serif',
+    fontSize: '18px',
+  },
+  addNewTodo: {
+    width: '50%',
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    margin: '10px 0',
+    '& button': {
+      fontWeight: 500,
+      fontSize: '1em',
+      height: '2em',
+    },
+  },
+}));
+export default useStyles;
